requirements.txt
TwitterAPI==2.3.3
picamera
rpio
